requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==20.1.0